feat: add status messages, check if data is complete, refactor codap utils [PT-188274657]
