+++ src/components/location-tab.scss
@@ -155,10 +155,48 @@
 
   .data-buttons.plugin-row {
     margin-top: -5px;
-    justify-content: flex-end;
+    justify-content: space-between;
 
-    button {
-      margin-left: 10px;
+    .button-container {
+      display: flex;
+      justify-content: flex-end;
+      flex-direction: row;
+
+      button {
+        margin-left: 10px;
+        white-space: nowrap;
+      }
+    }
+
+    .app-message div {
+      color: $teal-dark;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      font-size: 11.5px;
+
+      svg {
+        margin-right: 5px;
+      }
+
+      &.progress-indicator {
+        @keyframes rotateClockwise {
+          0% {
+            transform: rotate(0deg);
+          }
+          100% {
+            transform: rotate(360deg);
+          }
+        }
+
+        svg {
+          animation: rotateClockwise 1s linear infinite;
+        }
+      }
+
+      &.done, &.incomplete, &.error {
+        font-weight: 600;
+      }
     }
   }
